requirements.txt
fastapi[standard]
ruff
nanoid
supabase
python-dotenv
